refactor: remoção de customers e ajustes de tipos e parâmetros

+++ app/models/price_filter_model.py
@@ -3,8 +3,8 @@ from app.models import QueryModel
 
 
 class PriceFilter(QueryModel):
-    de__lt: Optional[float] = None
-    de__gt: Optional[float] = None
-    por__lt: Optional[float] = None
-    por__gt: Optional[float] = None
+    de__lt: Optional[int] = None
+    de__gt: Optional[int] = None
+    por__lt: Optional[int] = None
+    por__gt: Optional[int] = None
     sku: Optional[str] = None
